add the error save management code into a manager class
